requirements.txt
mysql-connector-python
sqlalchemy
fastapi
uvicorn[standard]
pyJWT[crypto]